FL-78 Microsd test application (#230)
* variable cluster size, label info functions for targets
* app record
* init, mount, get label, get sn, get space
* enable exfat
* more stack for the stack god
* remove c app and add cpp app
* remove MULTI_PARTITION
* fix 4gb bug
* update app to new template lib, add animated waiting
* tiny buffer configuration
* write speed benchmark
* fnv1a hash library
* make DEFAULT_STACK_SIZE and MAX_TASK_COUNT defined per target
* fully functional sd card app
* build sd test app to release firmware
* cpp, not c
* light up red led if error
* flags for c++
* linking with g++
* suppres snprintf warning
* move format work area to heap

+++ firmware/targets/f2/Src/fatfs/stm32_adafruit_sd.h
@@ -161,7 +161,7 @@ typedef struct
 {
   SD_CSD Csd;
   SD_CID Cid;
-  uint32_t CardCapacity;              /*!< Card Capacity */
+  uint64_t CardCapacity;              /*!< Card Capacity */
   uint32_t CardBlockSize;             /*!< Card Block Size */
   uint32_t LogBlockNbr;               /*!< Specifies the Card logical Capacity in blocks   */
   uint32_t LogBlockSize;              /*!< Specifies logical block size in bytes           */
